+++ warehouse/integrations/snowflake/datatype_mapper_test.go
@@ -1,0 +1,27 @@
+package snowflake
+
+import (
+	"database/sql"
+	"github.com/stretchr/testify/require"
+	"testing"
+)
+
+func TestCalculateDataType(t *testing.T) {
+	testCases := []struct {
+		columnType   string
+		numericScale sql.NullInt64
+		expected     string
+		exists       bool
+	}{
+		{"VARCHAR", sql.NullInt64{}, "string", true},
+		{"INT", sql.NullInt64{}, "int", true},
+		{"INT", sql.NullInt64{Int64: 2, Valid: true}, "float", true},
+		{"UNKNOWN", sql.NullInt64{}, "", false},
+	}
+
+	for _, tc := range testCases {
+		dataType, exists := calculateDataType(tc.columnType, tc.numericScale)
+		require.Equal(t, tc.expected, dataType)
+		require.Equal(t, tc.exists, exists)
+	}
+}
