+++ warehouse/integrations/snowflake/snowflake.go
@@ -43,48 +43,6 @@ const (
 	Application        = "Rudderstack_Warehouse"
 )
 
-var dataTypesMap = map[string]string{
-	"boolean":  "boolean",
-	"int":      "number",
-	"bigint":   "number",
-	"float":    "double precision",
-	"string":   "varchar",
-	"datetime": "timestamp_tz",
-	"json":     "variant",
-}
-
-var dataTypesMapToRudder = map[string]string{
-	"NUMBER":           "int",
-	"DECIMAL":          "int",
-	"NUMERIC":          "int",
-	"INT":              "int",
-	"INTEGER":          "int",
-	"BIGINT":           "int",
-	"SMALLINT":         "int",
-	"FLOAT":            "float",
-	"FLOAT4":           "float",
-	"FLOAT8":           "float",
-	"DOUBLE":           "float",
-	"REAL":             "float",
-	"DOUBLE PRECISION": "float",
-	"BOOLEAN":          "boolean",
-	"TEXT":             "string",
-	"VARCHAR":          "string",
-	"CHAR":             "string",
-	"CHARACTER":        "string",
-	"STRING":           "string",
-	"BINARY":           "string",
-	"VARBINARY":        "string",
-	"TIMESTAMP_NTZ":    "datetime",
-	"DATE":             "datetime",
-	"DATETIME":         "datetime",
-	"TIME":             "datetime",
-	"TIMESTAMP":        "datetime",
-	"TIMESTAMP_LTZ":    "datetime",
-	"TIMESTAMP_TZ":     "datetime",
-	"VARIANT":          "json",
-}
-
 var primaryKeyMap = map[string]string{
 	usersTable:      "ID",
 	identifiesTable: "ID",
@@ -1340,16 +1298,6 @@ func (sf *Snowflake) FetchSchema(model.Warehouse) (model.Schema, model.Schema, e
 	}
 
 	return schema, unrecognizedSchema, nil
-}
-
-func calculateDataType(columnType string, numericScale sql.NullInt64) (string, bool) {
-	if datatype, ok := dataTypesMapToRudder[columnType]; ok {
-		if datatype == "int" && numericScale.Valid && numericScale.Int64 > 0 {
-			datatype = "float"
-		}
-		return datatype, true
-	}
-	return "", false
 }
 
 func (sf *Snowflake) Cleanup() {
